# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NGOConnect AI resource coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Error"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Error"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      security:
        - bearerAuth: []
      parameters:
        - name: direction
          in: query
          schema:
            type: string
            enum: [incoming, outgoing, all]
            default: all
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated request list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestPage"
    post:
      operationId: createRequest
      tags: [requests]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceRequest"
  /requests/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: getRequest
      tags: [requests]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceRequest"
  /requests/{id}/accept:
    post:
      operationId: acceptRequest
      tags: [requests]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Accepted request and transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /requests/{id}/reject:
    post:
      operationId: rejectRequest
      tags: [requests]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Rejected request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceRequest"
  /requests/{id}/complete:
    post:
      operationId: completeRequest
      tags: [requests]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Completed transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        name: { type: string }
        role:
          type: string
          enum: [ngo, donor, admin]
        ngoId: { type: ["integer", "null"] }
    RegisterInput:
      type: object
      required: [email, password, name, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string, minLength: 2 }
        role:
          type: string
          enum: [ngo, donor]
        ngoName: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
...
    ResourceRequest:
      type: object
      required: [id, requesterNgoId, requesterNgoName, resourceType, quantity, urgency, location, requiredDate, status, createdAt]
      properties:
        id: { type: integer }
        requesterNgoId: { type: integer }
        requesterNgoName: { type: string }
        resourceType: { type: string }
        category:
          type: string
          enum: [food, medicines, healthcare_supplies, clothing, other]
        quantity: { type: number }
        urgency:
          type: string
          enum: [low, medium, high, critical]
        location: { type: string }
        requiredDate: { type: string, format: date }
        status:
          type: string
          enum: [open, accepted, rejected, completed, cancelled]
        createdAt: { type: string, format: date-time }
    RequestInput:
      type: object
      required: [resourceType, quantity, urgency, location, requiredDate]
      properties:
        resourceType: { type: string, minLength: 2 }
        category:
          type: string
          enum: [food, medicines, healthcare_supplies, clothing, other]
        quantity: { type: number, exclusiveMinimum: 0 }
        urgency:
          type: string
          enum: [low, medium, high, critical]
        location: { type: string, minLength: 2 }
        requiredDate: { type: string, format: date }
    Transaction:
      type: object
      required: [id, requestId, resourceId, quantity, status, createdAt]
      properties:
        id: { type: integer }
        requestId: { type: integer }
        resourceId: { type: integer }
        quantity: { type: number }
        status:
          type: string
          enum: [pending, accepted, completed, cancelled]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    RequestPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ResourceRequest" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
    Status:
      name: status
      in: query
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT